requirements.txt
# Gestion des variables d'environnement
python-dotenv==1.0.0

# Nécessaire pour l'API Discord
discord.py==2.3.2
aiohttp>=3.7.4,<4
audioop-lts; python_version>='3.13'

# API Twitch
twitchAPI>=4.5.0

# Nécessaire pour l'hébergement du site web
flask>=2.3.2
flask-login>=0.6.3
flask-sqlalchemy>=3.0.3
flask[async]
waitress>=3.0.2

# Nécessaire pour l'appel à l'API Humble Bundle
requests>=2.32.4

# Nécessaire pour l'appel à ProtonDB
algoliasearch>=4,<5

